pyproject.toml
[project]
name = "understanding-context-in-agentic-ai-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.0",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
]
